Finished the main code
Finished the login functionality. Now the code will check if your username or password is right and it will show you the right error message depending on which one is wrong. Created a global.php file to store things that need to be repeated often. Cleaned up the code a bit